Moved fetstStaff function to StaffContext, added refetch on successful user add and added snackbar rendering for some successful responses
