+++ nextjs/src/context/StaffContext.tsx
@@ -2,6 +2,7 @@
 import { createContext, useContext, Dispatch, SetStateAction, useState, PropsWithChildren } from "react";
 import { useSearchParams } from "next/navigation";
 import { useEffect } from "react";
+import axios from "@/axiosInstance";
 
 export enum Gender {
   Male = "male",
@@ -44,6 +45,7 @@ interface FilterContextProps {
   setLoading: Dispatch<SetStateAction<boolean>>;
   totalPages: number;
   setTotalPages: Dispatch<SetStateAction<number>>;
+  fetchStaff: (studentStaffOnly?: boolean) => void;
 }
 
 const StaffContext = createContext<FilterContextProps>({
@@ -55,6 +57,7 @@ const StaffContext = createContext<FilterContextProps>({
   setLoading: () => {},
   totalPages: 1,
   setTotalPages: () => {},
+  fetchStaff: () => {},
 });
 
 export const StaffContextProvider = ({ children }: PropsWithChildren) => {
@@ -78,8 +81,42 @@ export const StaffContextProvider = ({ children }: PropsWithChildren) => {
     }
   }, [searchParams]);
 
+  function fetchStaff(studentStaffOnly?: boolean) {
+    setStaff([]);
+    setLoading(true);
+    axios
+      .get("/student_staff", {
+        params: {
+          ...filter,
+          studentStaffOnly,
+        },
+      })
+      .then((res) => {
+        const mappedStaff: StaffCardProps[] = res.data.results.map((staff: any) => ({
+          firstName: staff.firstName,
+          lastName: staff.lastName,
+          shifts: staff.shifts,
+          tasksCompleted: staff.tasksCompleted,
+          email: staff.email,
+          picture: staff.picture,
+          gender: staff.gender,
+          phone: staff.phone,
+        }));
+
+        setStaff(mappedStaff);
+        setTotalPages(res.data.pages);
+        setLoading(false);
+      })
+      .catch((err) => {
+        console.error(err);
+        setLoading(false);
+      });
+  }
+
   return (
-    <StaffContext.Provider value={{ setFilter, filter, Staff, setStaff, loading, setLoading, totalPages, setTotalPages }}>{children}</StaffContext.Provider>
+    <StaffContext.Provider value={{ setFilter, filter, Staff, setStaff, loading, setLoading, totalPages, setTotalPages, fetchStaff }}>
+      {children}
+    </StaffContext.Provider>
   );
 };
